AgenticSyncKit/requirements-agentic.txt
PyYAML>=6.0.1
networkx>=3.3
GitPython>=3.1.43
rich>=13.7.1
jsonschema>=4.22.0
